requirements.txt
deepspeed
rotary_embedding_torch
transformers==4.31.0
inflect
einops==0.4.1
unidecode
scipy
librosa==0.9.1
numba
torchaudio
